requirements.txt
fastapi==0.68.0
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
websockets==9.1
celery
sqlalchemy
